pyproject.toml
[tool.poetry]
name = "anki-sync"
version = "0.1.0"
description = "Syncs words from Google Sheets to Anki notes."
authors = ["Your Name <you@example.com>"] # Please update with your details!
readme = "README.md"
packages = [{include = "anki_sync"}]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.0"
genanki = "^0.13.0"
google-api-python-client = "^2.0"
google-auth-httplib2 = "^0.1"
google-auth-oauthlib = "^0.5"
google-cloud-texttospeech = "^2.27.0"
pydantic = "^2.0"
elevenlabs = "^2.1.0"
pandas = "^2.2.0"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
google-genai = "^1.20.0"
attrs = "^25.3.0"
ankipandas = "^0.3.15"
modern-greek-inflexion = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
ruff = "^0.1.0"
ipython = "^9.3.0"
autoflake = "^2.3.1"
isort = "^6.0.1"
black = "^25.1.0"

[tool.poetry.scripts]
anki-sync = "anki_sync.cli:main"

[tool.flake8]
exclude = ['.git','__pycache__','.venv']
max-line-length = 120
application-import-names = 'rmx_edit_showcase'
extend-ignore = ['E722','E402']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
